+++ mobile/android/android-components/components/support/ktx/src/main/java/mozilla/components/support/ktx/android/view/Window.kt
@@ -6,8 +6,16 @@ package mozilla.components.support.ktx.android.view
 
 import android.os.Build
 import android.os.Build.VERSION.SDK_INT
+import android.os.Build.VERSION_CODES
+import android.view.View
 import android.view.Window
+import android.view.WindowManager
 import androidx.annotation.ColorInt
+import androidx.core.graphics.Insets
+import androidx.core.view.ViewCompat
+import androidx.core.view.WindowCompat
+import androidx.core.view.WindowInsetsCompat.Type.displayCutout
+import androidx.core.view.WindowInsetsCompat.Type.systemBars
 import androidx.core.view.WindowInsetsControllerCompat
 import mozilla.components.support.utils.ColorUtils.isDark
 
@@ -75,5 +83,42 @@ fun Window.setNavigationBarDividerColorCompat(@ColorInt color: Int?) {
     if (SDK_INT >= Build.VERSION_CODES.P && SDK_INT < Build.VERSION_CODES.VANILLA_ICE_CREAM) {
         @Suppress("DEPRECATION")
         navigationBarDividerColor = color ?: 0
+    }
+}
+
+/**
+ * Setup handling persistent insets - system bars and display cutouts ourselves instead of the framework.
+ * This results in keeping the same behavior for such insets while allowing to separately control the behavior
+ * for other dynamic insets.
+ *
+ * This only works on Android Q and above. On older versions calling this will result in no-op.
+ */
+fun Window.setupPersistentInsets() {
+    if (SDK_INT >= VERSION_CODES.Q) {
+        WindowCompat.setDecorFitsSystemWindows(this, false)
+
+        val rootView = decorView.findViewById<View>(android.R.id.content)
+        val persistentInsetsTypes = systemBars() or displayCutout()
+
+        ViewCompat.setOnApplyWindowInsetsListener(rootView) { _, windowInsets ->
+            val isInImmersiveMode = attributes.flags and WindowManager.LayoutParams.FLAG_LAYOUT_NO_LIMITS != 0
+            val persistentInsets = when (isInImmersiveMode) {
+                true -> {
+                    // If we are in immersive mode we need to reset current paddings and avoid setting others.
+                    Insets.of(0, 0, 0, 0)
+                }
+                false -> windowInsets.getInsets(persistentInsetsTypes)
+            }
+
+            rootView.setPadding(
+                persistentInsets.left,
+                persistentInsets.top,
+                persistentInsets.right,
+                persistentInsets.bottom,
+            )
+
+            // Pass window insets further to allow below listeners also know when there is a change.
+            return@setOnApplyWindowInsetsListener windowInsets
+        }
     }
 }